backend/requirements.txt
# ── Backend Dependencies ──
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Celery
celery[redis]==5.3.6
redis==5.0.1

# Storage
minio==7.2.3
boto3==1.34.34

# Analysis
yara-python==4.5.0
python-magic==0.4.27
pefile==2023.2.7
lief==0.14.1

# Validation & serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# Utilities
httpx==0.26.0
structlog==24.1.0
python-dateutil==2.8.2
orjson==3.9.13
